requirements.txt
sqlalchemy>=2.0.30
psycopg[binary]>=3.2
python-dotenv>=1.0
alembic>=1.13
chromadb>=0.5.8
sentence-transformers>=3.0.1
openai>=1.45.0
streamlit>=1.39.0
